@@ -6,34 +6,35 @@ package mocks
 
 import (
 	context "context"
-	gomock "github.com/golang/mock/gomock"
 	reflect "reflect"
+
+	gomock "github.com/golang/mock/gomock"
 )
 
-// MockFilesFinder is a mock of FilesFinder interface
+// MockFilesFinder is a mock of FilesFinder interface.
 type MockFilesFinder struct {
 	ctrl     *gomock.Controller
 	recorder *MockFilesFinderMockRecorder
 }
 
-// MockFilesFinderMockRecorder is the mock recorder for MockFilesFinder
+// MockFilesFinderMockRecorder is the mock recorder for MockFilesFinder.
 type MockFilesFinderMockRecorder struct {
 	mock *MockFilesFinder
 }
 
-// NewMockFilesFinder creates a new mock instance
+// NewMockFilesFinder creates a new mock instance.
 func NewMockFilesFinder(ctrl *gomock.Controller) *MockFilesFinder {
 	mock := &MockFilesFinder{ctrl: ctrl}
 	mock.recorder = &MockFilesFinderMockRecorder{mock}
 	return mock
 }
 
-// EXPECT returns an object that allows the caller to indicate expected use
+// EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockFilesFinder) EXPECT() *MockFilesFinderMockRecorder {
 	return m.recorder
 }
 
-// Run mocks base method
+// Run mocks base method.
 func (m *MockFilesFinder) Run(arg0 context.Context, arg1 chan<- string) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Run", arg0, arg1)
@@ -41,7 +42,7 @@ func (m *MockFilesFinder) Run(arg0 context.Context, arg1 chan<- string) error {
 	return ret0
 }
 
-// Run indicates an expected call of Run
+// Run indicates an expected call of Run.
 func (mr *MockFilesFinderMockRecorder) Run(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Run", reflect.TypeOf((*MockFilesFinder)(nil).Run), arg0, arg1)
